make all test in bid work and refactor some par of bidController

+++ test/bid.js
@@ -2,6 +2,7 @@ var chai = require('chai')
 var chaiHttp = require('chai-http')
 require('dotenv').config()
 let app = require('../app')
+let _ = require('lodash')
 
 var should = chai.should()
 chai.use(chaiHttp)
@@ -15,39 +16,53 @@ describe('Bid Test', () => {
   describe('Bidding', () => {
     it('Should be return status Success when trying to bid a auction', (done) => {
       models.Auction.findAll().then(auctions => {
-        let auctionsLength = auctions.length
+        let convertedAuctions = JSON.parse(JSON.stringify(auctions))
+        let auctionsLength = convertedAuctions.length
         if (auctionsLength != 0) {
-          // console.log('ambil auction terakhir : ', auctions);
-          // console.log('id terakhir ? ', auctions[auctionsLength - 1].id);
-          chai.request(serverHost).post('/bids/').send({
-            userId: 1,
-            token: process.env.BUKALAPAK_TOKEN,
-            auctionId: auctions[auctionsLength - 1].id,
-            nextBid: 140000
-          }).end((err, res) => {
-            if (err) {
-              done(err)
-            } else {
-              // console.log('isi res di test : ', res.body);
-              res.should.have.status(200);
-              res.should.be.json;
-              res.body.success.should.to.equal(true)
-              // delete bid after test
-              models.Bid.findById(res.body.id).then(bid => {
-                if (bid) {
-                  bid.destroy({
-                    where: {
-                      id : bid.id
-                    }
-                  }).then(() => {
-                    console.log('delete bid after test success');
-                    done()
+          // find the highest bid for last auction
+          let lastAuction = convertedAuctions[auctionsLength - 1]
+          models.Bid.findAll({
+            where: {
+              auctionId: lastAuction.id
+            }
+          }).then(bids => {
+            if (bids.length > 0) {
+              let highestBid = _.maxBy(JSON.parse(JSON.stringify(bids)), 'current_bid')
+              let nextBid = highestBid.current_bid + lastAuction.kelipatan_bid
+              // console.log('ambil auction terakhir : ', convertedAuctions);
+              // console.log('id terakhir ? ', convertedAuctions[auctionsLength - 1].id);
+              chai.request(serverHost).post('/bids/').send({
+                userId: 1,
+                token: process.env.BUKALAPAK_TOKEN,
+                auctionId: lastAuction.id,
+                nextBid: nextBid
+              }).end((err, res) => {
+                if (err) {
+                  done(err)
+                } else {
+                  // console.log('isi res di test : ', res.body);
+                  res.should.have.status(200);
+                  res.should.be.json;
+                  res.body.success.should.to.equal(true)
+                  // delete bid after test
+                  models.Bid.findById(res.body.id).then(bid => {
+                    if (bid) {
+                      bid.destroy({
+                        where: {
+                          id : bid.id
+                        }
+                      }).then(() => {
+                        console.log('delete bid after test success');
+                        done()
 
+                      })
+                    }
                   })
                 }
-              })
+              });
             }
-          });
+          })
+
         } else {
           console.log('tidak ada auctions di DB');
         }
@@ -55,44 +70,61 @@ describe('Bid Test', () => {
     })
     it('Should be return all field / property when trying to bid a auction', (done) => {
       models.Auction.findAll().then(auctions => {
-        const auctionsLength = auctions.length
-        if (auctions.length != 0) {
-          // console.log('ambil auction terakhir : ', auctions);
-          // console.log('id terakhir ? ', auctions[auctionsLength - 1].id);
-          chai.request(serverHost).post('/bids/').send({
-            userId: 1,
-            token: process.env.BUKALAPAK_TOKEN,
-            auctionId: auctions[auctionsLength - 1].id,
-            nextBid: 140000
-          }).end((err, res) => {
-            if (err) {
-              done(err)
-            } else {
-              res.should.have.status(200);
-              res.should.be.json;
-              res.body.success.should.to.equal(true)
-              res.body.should.have.property('id')
-              res.body.should.have.property('auctionId')
-              res.body.should.have.property('current_price')
-              res.body.should.have.property('minimum_next_bidding')
-              res.body.should.have.property('success')
-              res.body.should.have.property('message')
-              // delete bid after test
-              models.Bid.findById(res.body.id).then(bid => {
-                if (bid) {
-                  bid.destroy({
-                    where: {
-                      id : bid.id
-                    }
-                  }).then(() => {
-                    console.log('delete bid after test success');
-                    done()
+        let convertedAuctions = JSON.parse(JSON.stringify(auctions))
+        let auctionsLength = convertedAuctions.length
+        if (auctionsLength != 0) {
+          // find the highest bid for last auction
+          let lastAuction = convertedAuctions[auctionsLength - 1]
+          models.Bid.findAll({
+            where: {
+              auctionId: lastAuction.id
+            }
+          }).then(bids => {
+            if (bids.length > 0) {
+              let highestBid = _.maxBy(JSON.parse(JSON.stringify(bids)), 'current_bid')
+              let nextBid = highestBid.current_bid + lastAuction.kelipatan_bid
+              // console.log('ambil auction terakhir : ', convertedAuctions);
+              // console.log('id terakhir ? ', convertedAuctions[auctionsLength - 1].id);
+              chai.request(serverHost).post('/bids/').send({
+                userId: 1,
+                token: process.env.BUKALAPAK_TOKEN,
+                auctionId: lastAuction.id,
+                nextBid: nextBid
+              }).end((err, res) => {
+                if (err) {
+                  done(err)
+                } else {
+                  // console.log('isi res di test : ', res.body);
+                  res.should.have.status(200);
+                  res.should.be.json;
+                  res.body.should.have.property('id')
+                  res.body.should.have.property('auctionId')
+                  res.body.should.have.property('username')
+                  res.body.should.have.property('name')
+                  res.body.should.have.property('bidding_time')
+                  res.body.should.have.property('current_price')
+                  res.body.should.have.property('minimum_next_bidding')
+                  res.body.should.have.property('success')
+                  res.body.should.have.property('message')
+                  // delete bid after test
+                  models.Bid.findById(res.body.id).then(bid => {
+                    if (bid) {
+                      bid.destroy({
+                        where: {
+                          id : bid.id
+                        }
+                      }).then(() => {
+                        console.log('delete bid after test success');
+                        done()
 
+                      })
+                    }
                   })
                 }
-              })
+              });
             }
-          });
+          })
+
         } else {
           console.log('tidak ada auctions di DB');
         }
